Fix linter2
