+++ .eslintrc.json
@@ -15,6 +15,8 @@
     "extends": ["airbnb", "plugin:react/recommended"],
     "plugins": ["react"],
     "rules": {
+      "jsx-a11y/label-has-associated-control": "off",
+      "jsx-a11y/label-has-for":"off",
       "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
       "react/react-in-jsx-scope": "off",
       "import/no-unresolved": "off",
