--- src/components/Book.jsx
@@ -1,51 +1,0 @@
-const Book = () => (
-    <>
-      <div className="card">
-        <div className="title">
-          <span>Action</span>
-          <h1>The hunger games</h1>
-          <p>Unseen Content</p>
-          <div class="action-buttons"><button class="button-outline" type="button">Comments</button>
-          <div class="vertical-divider"></div>
-          <button class="button-outline" type="button">Remove</button>
-          <div class="vertical-divider"></div>
-          <button class="button-outline" type="button">Edit</button>
-          </div> 
-          </div>
-
-        <div className="progress">
-          <h4>60%</h4>
-        </div>
-        <div className="chapter">
-          <h4>Current Chapter</h4>
-          <h5>Chapter 16</h5>
-          <button type="button">UPDATE PROGRESS</button>
-        </div>
-      </div>
-  
-      <div className="card">
-        <div className="title">
-          <span>Action</span>
-          <h1>The hunger games</h1>
-          <p>Unseen Content</p>
-          <div class="action-buttons"><button class="button-outline" type="button">Comments</button>
-          <div class="vertical-divider"></div>
-          <button class="button-outline" type="button">Remove</button>
-          <div class="vertical-divider"></div>
-          <button class="button-outline" type="button">Edit</button>
-          </div> 
-        </div>
-  
-        <div className="progress">
-          <h4>60%</h4>
-        </div>
-        <div className="chapter">
-          <h4>Current Chapter</h4>
-          <h5>Chapter 16</h5>
-          <button type="button">UPDATE PROGRESS</button>
-        </div>
-      </div>
-    </>
-  );
-  
-  export default Book;
