+++ src/components/Books.js
@@ -1,0 +1,11 @@
+import Form from './Form';
+import Book from './Book';
+
+const Books = () => (
+  <div>
+    <Book />
+    <Form />
+  </div>
+);
+
+export default Books;
