--- src/components/Books.jsx
@@ -1,11 +1,0 @@
-import Form from './Form.jsx';
-import Book from './Book.jsx';
-
-const Books = () => (
-  <div>
-    <Book />
-    <Form />
-  </div>
-);
-
-export default Books;
