+++ src/components/Categories.js
@@ -1,0 +1,7 @@
+const Category = () => (
+  <div>
+    <button type="button">Check status</button>
+  </div>
+);
+
+export default Category;
