--- src/components/Categories.jsx
@@ -1,7 +1,0 @@
-const Category = () => (
-    <div>
-      <button type="button">Check status</button>
-    </div>
-  );
-  
-  export default Category;
